+++ include/constants/map_groups.h
@@ -816,6 +816,9 @@
 #define MAP_FALADOR_CASTLE_3F        (16 | (42 << 8))
 #define MAP_FALADOR_MACE_STORE       (17 | (42 << 8))
 #define MAP_FALADOR_PARTY_ROOM       (18 | (42 << 8))
+#define MAP_FALADOR_CASTLE_1F_R2     (19 | (42 << 8))
+#define MAP_FALADOR_CASTLE_1F_R3     (20 | (42 << 8))
+#define MAP_FALADOR_CASTLE_3F_R2     (21 | (42 << 8))
 
 // gMapGroup_PokescapeIndoorVarrock
 #define MAP_VARROCK_THESSALIA_SHOP    (0 | (43 << 8))